Made locations open > 1 hour a solid green dot
Description: Locations that will remain open for more than an hour now use a solid green dot. Locations remaining open for an hour or less remain blinking. Please see "Potential additional feature."
Closes #2
Potential additional feature: An additional feature that should be very easy to implement if necessary is making red dots blink if the location opens in an hour or less. This feature has a similar implementation to the blinking green dot, and should use start instead of end as well as slightly change the math to calculate timeDiff for when the location is closed.