Updated map view
CMU Concept restaurants that close soon are now yellow on the map, and the rest are red and green as usual, depending on if they are open or closed. This way, the list view matches the map. Before, restaurants that appeared yellow on the list were green on the map.